# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ChatMe - WhatsApp-like real-time messaging API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface LoginInput {
  /**
     * @minLength 2
     * @maxLength 30
     */
  username: string;
  displayName?: string;
}

export interface User {
  id: number;
  username: string;
  displayName: string;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  statusText?: string | null;
  isOnline?: boolean;
  /** @nullable */
  lastSeen?: string | null;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface ProfileUpdate {
  displayName?: string;
  avatarUrl?: string;
  statusText?: string;
}

export type ConversationType = typeof ConversationType[keyof typeof ConversationType];


export const ConversationType = {
  direct: 'direct',
  group: 'group',
} as const;

export type ConversationMemberRole = typeof ConversationMemberRole[keyof typeof ConversationMemberRole];


export const ConversationMemberRole = {
  admin: 'admin',
  member: 'member',
} as const;

export interface ConversationMember {
  userId: number;
  role: ConversationMemberRole;
  joinedAt: string;
  user?: User;
}

export interface Conversation {
  id: number;
  type: ConversationType;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  createdBy?: number | null;
  createdAt: string;
  members: ConversationMember[];
}

export type ConversationSummaryType = typeof ConversationSummaryType[keyof typeof ConversationSummaryType];


export const ConversationSummaryType = {
  direct: 'direct',
  group: 'group',
} as const;

export type MessageType = typeof MessageType[keyof typeof MessageType];


export const MessageType = {
  text: 'text',
  image: 'image',
  video: 'video',
  file: 'file',
  voice: 'voice',
  system: 'system',
} as const;

export interface MessagePreview {
  id: number;
  /** @nullable */
  content: string | null;
  senderId: number;
  senderName?: string;
  type?: string;
}

export interface Reaction {
  emoji: string;
  userId: number;
  userName?: string;
}

export interface Message {
  id: number;
  conversationId: number;
  senderId: number;
  sender?: User;
  type: MessageType;
  /** @nullable */
  content?: string | null;
  /** @nullable */
  mediaUrl?: string | null;
  /** @nullable */
  mediaType?: string | null;
  /** @nullable */
  mediaSize?: number | null;
  /** @nullable */
  fileName?: string | null;
  /** @nullable */
  replyToId?: number | null;
  replyTo?: MessagePreview;
  reactions?: Reaction[];
  readBy?: number[];
  isDeleted: boolean;
  createdAt: string;
  updatedAt?: string;
}

export interface ConversationSummary {
  id: number;
  type: ConversationSummaryType;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  createdBy?: number | null;
  createdAt: string;
  lastMessage?: Message;
  unreadCount?: number;
  otherUser?: User;
  memberCount?: number;
}

export type ConversationInputType = typeof ConversationInputType[keyof typeof ConversationInputType];


export const ConversationInputType = {
  direct: 'direct',
  group: 'group',
} as const;

export interface ConversationInput {
  type: ConversationInputType;
  memberIds?: number[];
  name?: string;
  description?: string;
  avatarUrl?: string;
}

export interface ConversationUpdate {
  name?: string;
  description?: string;
  avatarUrl?: string;
}

export interface AddMembersInput {
  userIds: number[];
}

export interface UnreadCount {
  conversationId: number;
  count: number;
}

export type MessageInputType = typeof MessageInputType[keyof typeof MessageInputType];


export const MessageInputType = {
  text: 'text',
  image: 'image',
  video: 'video',
  file: 'file',
  voice: 'voice',
} as const;

export interface MessageInput {
  type: MessageInputType;
  content?: string;
  mediaUrl?: string;
  mediaType?: string;
  mediaSize?: number;
  fileName?: string;
  replyToId?: number;
}

export interface ReactionInput {
  emoji: string;
}

export interface FileUploadResponse {
  url: string;
  type: string;
  size: number;
  fileName?: string;
}

export type ListUsersParams = {
/**
 * Search by username or display name
 */
search?: string;
};

export type ListMessagesParams = {
/**
 * Load messages before this message ID (cursor)
 */
before?: number;
limit?: number;
};

export type SearchMessagesParams = {
q: string;
conversationId?: number;
};